NB1: Remove A1N support
It is better to split the repos rather than having a single one.
Nokia 8 launched with Nougat and uses software encryption, so any Sirocco blobs will not work for it.
Even though NB1 blobs will probably decrypt on A1N, it is better to use the hardware based keymaster on A1N